removed mysql import statement in database utilities

+++ database/database_util.py
@@ -1,6 +1,7 @@
 from sql.schema import *
 from database.data_extractors import *
-from database.database_util_mysql import DBMySQL
+
+# from database.database_util_mysql import DBMySQL
 from database.database_util_postgres import DBPostgreSQL
 
 
@@ -13,16 +14,16 @@ def get_db_object(db, noisy=False):
     :param noisy: if True, print database being used
     :return: BaseDB object
     """
-    if db == 'postgres':
+    if db == "postgres":
         if noisy:
             print(f"USING POSTGRES {DB_KEYS_POSTGRES['dbname']}")
         _db = DBPostgreSQL(DB_KEYS_POSTGRES)
-    elif db == 'mysql':
+    elif db == "mysql":
         if noisy:
             print(f"USING MYSQL DATABASE {DB_KEYS_MYSQL['database']}")
         _db = DBMySQL(DB_KEYS_MYSQL)
     else:
-        raise ValueError(f'unknown database {db}')
+        raise ValueError(f"unknown database {db}")
     return _db
 
 
@@ -33,7 +34,7 @@ def store_release_metadata(db, release):
     :param release: discogs_client.Release object
     :return: None
     """
-    print(f'Storing release metadata for: {release.title} by {release.artists[0].name}')
+    print(f"Storing release metadata for: {release.title} by {release.artists[0].name}")
 
     release_info = prepare_release_data(release)
     db.insert_rows(release_info, RELEASE_TABLE)
@@ -57,8 +58,8 @@ def store_release_data(release, store_metadata=True, store_prices=True, db=DB_CH
     :param db: str, which database to use (mysql or postgres)
     :return: None
     """
-    assert isinstance(release, discogs_client.Release), f'release is {type(release)}'
-    print(f'Storing marketplace data for: {release.title} by {release.artists[0].name}')
+    assert isinstance(release, discogs_client.Release), f"release is {type(release)}"
+    print(f"Storing marketplace data for: {release.title} by {release.artists[0].name}")
     _db = get_db_object(db)
     if store_prices:
         marketplace_data = prepare_price_data(release)
@@ -69,6 +70,7 @@ def store_release_data(release, store_metadata=True, store_prices=True, db=DB_CH
         except:
             pass
     return None
+
 
 #
 # -----------------------------------------------------------------------------
@@ -84,20 +86,20 @@ def get_price_data(db=DB_CHOICE, **conditions):
     """
     _db = get_db_object(db)
     df = _db.read_rows(PRICES_VIEW, **conditions)
-    df = df.sort_values(['release_id', 'when'])
+    df = df.sort_values(["release_id", "when"])
     # print(df.columns)
-    df['country'].fillna('-', inplace=True)
-    df['lowest_price'] = df['lowest_price'].astype(float)
+    df["country"].fillna("-", inplace=True)
+    df["lowest_price"] = df["lowest_price"].astype(float)
     return df
 
 
 def get_metadata(entity, db=DB_CHOICE, **conditions):
     _db = get_db_object(db)
-    if entity == 'label':
+    if entity == "label":
         output = _db.read_rows(LABEL_TABLE, **conditions)
-    elif entity == 'artist':
+    elif entity == "artist":
         output = _db.read_rows(ARTIST_TABLE, **conditions)
-    elif entity == 'album':
+    elif entity == "album":
         output = _db.read_rows(RELEASE_TABLE, **conditions)
     else:
         output = _db.read_rows(entity, **conditions)
